test: splitting blockstream test cases + lint fixes
