+++ packages/sources/blockstream/test/integration/adapter.test.ts
@@ -1,0 +1,67 @@
+import { Requester } from '@chainlink/ea-bootstrap'
+import { assertSuccess, assertError } from '@chainlink/ea-test-helpers'
+import { AdapterRequest } from '@chainlink/types'
+import { makeExecute } from '../../src/adapter'
+
+describe('execute', () => {
+  const jobID = '1'
+  const execute = makeExecute()
+
+  describe('successful calls @integration', () => {
+    const requests = [
+      {
+        name: 'empty data',
+        testData: { data: {} },
+      },
+      {
+        name: 'id not supplied (endpoint height)',
+        testData: {
+          data: {
+            endpoint: 'height',
+          },
+        },
+      },
+      {
+        name: 'endpoint difficulty',
+        testData: {
+          id: jobID,
+          data: {
+            endpoint: 'difficulty',
+          },
+        },
+      },
+    ]
+
+    requests.forEach((req) => {
+      it(`${req.name}`, async () => {
+        const data = await execute(req.testData as AdapterRequest)
+        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
+        expect(data.result).toBeGreaterThan(0)
+        expect(data.data.result).toBeGreaterThan(0)
+      })
+    })
+  })
+
+  describe('error calls @integration', () => {
+    const requests = [
+      {
+        name: 'unknown endpoint',
+        testData: {
+          id: jobID,
+          data: { endpoint: 'not_real' },
+        },
+      },
+    ]
+
+    requests.forEach((req) => {
+      it(`${req.name}`, async () => {
+        try {
+          await execute(req.testData as AdapterRequest)
+        } catch (error) {
+          const errorResp = Requester.errored(jobID, error)
+          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
+        }
+      })
+    })
+  })
+})
